+++ src/codeas/core/retriever.py
@@ -3,8 +3,8 @@ from typing import Optional
 
 from pydantic import BaseModel
 
-from codeag.core.metadata import RepoMetadata
-from codeag.ui.state import state
+from codeas.core.metadata import RepoMetadata
+from codeas.ui.state import state
 
 
 class ContextRetriever(BaseModel):
