+++ src/codeas/use_cases/documentation.py
@@ -1,9 +1,9 @@
-from codeag.core.agent import Agent
-from codeag.core.llm import LLMClient
-from codeag.core.metadata import RepoMetadata
-from codeag.core.repo import Repo
-from codeag.core.retriever import ContextRetriever
-from codeag.use_cases import prompts
+from codeas.core.agent import Agent
+from codeas.core.llm import LLMClient
+from codeas.core.metadata import RepoMetadata
+from codeas.core.repo import Repo
+from codeas.core.retriever import ContextRetriever
+from codeas.use_cases import prompts
 
 DOCS_MODEL = "gpt-4o"
 SECTION_CONFIG = {
@@ -78,8 +78,8 @@ def generate_docs_section(
 
 
 if __name__ == "__main__":
-    from codeag.core.metadata import RepoMetadata
-    from codeag.core.repo import Repo
+    from codeas.core.metadata import RepoMetadata
+    from codeas.core.repo import Repo
 
     llm_client = LLMClient()
     repo = Repo(repo_path=".")
